[refactor] DTO 올바른 패키지로 이동
